chore: move color updates to individual components and helper
