+++ apps/web/components/settings/design/1.fonts/types.ts
@@ -1,0 +1,1 @@
+export type FontSetting = { caption: string; value: string };
